fix a potential core-dump bug (only happening when 1d compression): fix method: predAbsErr<=checkRadius --> predAbsErr<checkRadius (previously, suppose there are 63 intervals, some predicted values with integer code 64 would be wrontly treated as 0, because 64==0 for unsigned)

+++ sz/src/DynamicDoubleArray.c
@@ -21,7 +21,7 @@ void new_DDA(DynamicDoubleArray **dda, size_t cap) {
 
 void convertDDAtoDoubles(DynamicDoubleArray *dba, double **data)
 {
-	int size = dba->size;
+	size_t size = dba->size;
 	if(size>0)
 		*data = (double*)malloc(size * sizeof(double));
 	else
